fix: update tokenizer to traverse the single-generate struct
We had been into()-ing the single-generate struct into a full Query, and then traversing the full query, with a few todos(). This eliminates the todos(), and also leverages the core Message enum.

This should pave the way to apply chat templates more cleanly -- as we now know (as we traverse) that we are in single-generate subtree. This is something assumed by the chat templates.

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>