fix: assistant msg cropping might crop suffix special tokens
The complexity here: we need to crop the assistant messages so that
they end on a block boundary. We cannot pad them out, because this is
not what the model server will do when generating the assistant
message. But it will cache the prefix of full blocks. Hence the need
to crop. However, we cannot just crop at the end, as this would also
crop off any "end of text" special tokens that the chat template adds
to the end of the `self.assistant(m)` token sequence. Therefore, we
need to crop just the message part. This logic tries to do all of that
in a way that is agnostic to the cast template. However, the logic
does currently assume that the chat template will never add special
tokens *in the middle* of the given message `m`; it assumes special
tokens are only ever added (if at all) to the beginning or end.

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>