test: when testing vllm-spnl speedup, test larger doc sizes
Updates rag builtin to support -k/--chunk-size to allow testing a variety of document sizes. As a result, the default document is no longer the PDL api spec.

Also updates the setup.sh driver to improve error handling. The use of `find` to loop over and execute the test scripts was not failing if any individual test script exited with failure.

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>